backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pydantic>=2.10.0
fastapi-mcp>=0.1.8
python-dotenv>=1.0.1
